refactor: split the generator engine into focused modules
src/mdx-to-nextjs-generator.ts shrinks from ~2760 lines to a thin
orchestrator that wires together eleven new modules under
src/generator/:

- secure-source-fs.ts: SecureSourceFs, a hardened fs layer that scopes
  reads to the documentation root and project root, guarding against
  path traversal and symlink escapes.
- app-scaffolder.ts: AppScaffolder writes the initial Next.js app
  structure (appStructure/obsoleteFiles/startingDocsStructure) plus
  the robots, next.config, pnpm-workspace, and proxy templates.
- project-config-repository.ts: ProjectConfigRepository reads,
  validates, and writes the JSON config files (navigation, sections,
  fonts, analytics) through SecureSourceFs.
- section-resolver.ts: validateSectionsConfig, discoverSections,
  addApiReferenceSection, and determineSectionRoute resolve an MDX
  file's section and slug.
- page-catalog.ts: parseMdxPageMeta, validateRouteCollisions, and
  mergePages build and merge the page-metadata catalog from MDX
  frontmatter.
- page-renderer.ts: renderMdxPage, renderHomepage, and
  renderSectionPage turn parsed MDX into page content and RSS route
  state.
- generated-route-manager.ts: GeneratedRouteManager reconciles routes
  owned by generated MDX pages via the GeneratedArtifacts registry.
- api-reference-generator.ts: ApiReferenceGenerator produces the
  OpenAPI-driven synthetic reference pages and routes.
- public-asset-manager.ts: PublicAssetManager mirrors the public/
  directory and manages the llms.txt/llms-full.txt/skill.md/
  .well-known/mcp.json aggregate files.
- site-artifacts.ts: buildSitemapEntries and related helpers generate
  sitemap.xml, robots.txt, and the llms files from the page catalog.
- watch-coordinator.ts: WatchCoordinator owns the chokidar watchers
  for MDX, config, font, analytics, doccupine.json, public/, and
  OpenAPI spec sources.

No behavioral changes are intended; this is a pure extraction to keep
each concern independently readable and testable.
